rework cache handling
If you have an item that should be unloaded on demand, you make that item have
a SwfdecCacheHandle and swfdec_cache_add_handle that handle to the cache just
before you load it. The cache will then automatically unload it when memory
gets sparse. If you unload the item yourself, use swfdec_cache_remove_handle on
it. See the gtk-doc documentation for those 2 functions for details.

What's missing? SwfdecSound should be loaded on demand, too. And SwfdecSound
and SwfdecImage should probably have a common ancestor (base class or interface)
for cached items.