make all scripts have a call object
This is required to be able to handle DefineLocal.
It will definitely break more functions than I've fixed here, so if you find
one that still assumes fp->scopeChain is the movie object, change it to use
fp->thisp.
Now all scripts have their own function, and they can access it via script->fun.
If the function gets GC'ed, it'll unset it again.