Do not ref the player anymore
Using a refcount was a good idea when SwfdecSource was seperate. Now it's
used from the player itself, so it'll get into a refcount loop.
We use weak references instead, so it's theoretically possible to use the
source outside of a player. It'll remove itself when the player goes away.