change swfdec_as_frame_find_variable() return value semantics
It used to look up the object that contained the variable.
Now it looks up the object in the scope chain that has the variable
in its prototype chain.
Consider this graphical piece of art (* denotes the object containing the variable,
scope chain is X axis, prototypes Y axis):
A  =>  B  =>  C
v      v
PA     PB*
       v
       PPB
Previously this function would return PB, now it returns B.