huge patch that reorganizes PlaceObject handling
Previously, the code was building up a range of frames for every piece of
content that showed when it was active. On backward gotos only those items were
kept where the frame was included in their range.
The new code just performs all actions again and copies old objects over where
appropriate. This produces resultss that are a lot closer to the official
player.