add has_color_transform and has_transform field to SwfdecContent
SwfdecContent needs to die, but that's another big reorg patch I'm scared of atm.
As far as I see it, a SwfdecSprite should just save all displaylist tags (the
stuff that's called actions right now) verbatim in a long list of buffers and
whenever we execute these tags, we just iterate through this list and parse the
tags.
This makes startup extremely fast (no need to parse anything) and fits exactly
into the curent model. And we save quite some meory as we don't need to allocate
loads of SwfdecContent structures.