fix setting undefined variables
This includes a rework of the target management. Every frame now contains
a ->target member that is always set to the current target. It also has an
->original_target member that remembers the original target, so a SetTarget
call can correctly reset it.
The previous notion of ->var_object has been removed.

An ->is_local property has been added. This porperty is used during
DefineLocal actions to decide wether to set proeprties on the current frame
or on the target. (FIXME: needs checking that those variables really are set on
the target)