rewrite stack handling
Excuse me for the big patch, but there is a lot of code that uses the stack.

So I changed the way the stack is handled. Previously every frame allocated a
private stack for 100 elements. This was neither very practicable (I have files
that require more stack space) nor very performant, since it required allocating
1200 bytes for every function call. This new method attempts to fix both of
these issues.

The stack is now allocated on the context, currently in segments of 8192 values
and is shared by all functions executed on the stack. If the stack space is
exceeded, a new segment is allocated. This seems to work ok, I'll write more
tests for this. The number for the stack size might also need to be tuned.

This of course has some implications. The biggest one is that doing weird stuff
on the stack (like popping values that are still needed, such as the often used
swfdec_as_value_to_integer (cx, swfdec_as_stack_pop (stack));) doesn't work
anymore since the stack can immediately be reused (when calling valueOf() in the
example above).
Another one is the handling of return values. Since the return value of
CallFunction is supposed to show up on the stack, it is now pushed only when
returning from the function call, not before calling the function, as it was
done before.

Another thing that got more complicated is that we can no longer assume that
arguments to a function can be treated as an array. In particular since it is
possible to call functions with more than 8192 arguments (I tested and Flash
calls functions with 10000000 arguments if I insist). So parsing these on a
function call got more complicated.

If you get any crashers and weird output regressions in the near future, please
git bisect the problem and if it was introduced by this commit, immediately
alert me about it. I won't claim this commit is bugfree. But it passes the
testsuite.