fix coordinate systems
Swfdec now has 4 types of coordinate systems:
- Stage (unique)
- global (unique)
- movie (one per movie)
- parent (one per movie)

The Stage coordinate system is what the user of the API sees and interacts with.
It's what coordinates in _handle_mouse() are given in and what the invalidate
signal uses for its rectangle. The Flash movie itself has no possibility to get
to this coordinate system.

The global coordinate system is the system all root movies live in. It's what
you get after applying the Stage transformations. The two functions
swfdec_player_stage_to_global() and swfdec_player_global_to_stage() transform
between this and the previous coordinate system. Various Flash functions (like
for example hitTest when using coordinates) expect coordinates from this system.

The there is one coordinate system per movie. It's what you get after applying
the movie's matrix to the parent's coordinate system. This coordinate system is
used for the _x, _y, _height and _width properties of the movie's children.
Drawing operations in the Flash file format are specified in this coordinate
system, too. The functions swfdec_movie_global_to_local() and
swfdec_movie_local_to_global() convert between a movie's and the global
coordinate systems. You can of course chain these functions with the stage
conversion functions to transform stage coordinates to movie coordinates and
vice versa.

Another coordinate system often used in Swfdec terminology is the parent system.
This is not actually a different coordinate system but just refers to the
coordinate system of a movie's parent or in the root movie case the global
coordinate system. Since quite some Actionscript variables and  functions
require coorinates in this system, it is useful to have a simple term for it.