add startup script for AS code
This works the same way as the normal init script in swfdec_initialize.as.
The only difference is that this script is for the stuff that should be in
SwfdecAsContext as opposed to the old one that was only for stuff that should
end up in SwfdecPlayer, but not in SwfdecAsContext.

This patch alse makes Number use that script as a proof-of-concept and bugfix.