rewrite scope chain handling
Previously we had lots of scope objects that were instantiated whenever the
scope chain changed.
Now it's just a list of objects in the scope chain.

It also adds a function swfdec_as_frame_push_block() that allows registering
a callback that gets called whenever the program counter leaves that block. It
is used for implementing With and should also work for try/catch.