rework input data handling.
The code now creates a buffer for both compressed and uncompressed Flash files.
This is the first step in fixing the script engine buffer size limitation, which
triggered quite often for "protected" Flash files.

Most of the obfuscators or code protectors for Flash files do not try to
obfuscate the bytecodes, but instead just move the code to a different part of
the file and use a jump bytecode to get there.
One undocumented feature of Flash is that the script engine itself allows jumps
to every part of the Flash file, not limited to the currently executing script.