rewrite GStreamer backend to directly use the element and not create a pipeline
That was Wim's suggested method, and it works fine. The only issue right now is
that GStreamer also does the resampling, and a) audioresample is slow and
b) the whole process takes a lot of time. This causes the app to stop for half a
second when a long sample gets decoded. Not good.