remove commented out and non-working code for native FLV playback
The code had too many hacks and special cases, and that's one thing I don't like.
Especially ot for an add-on feature.
If we ever introduce native FLV playback again, we should do this via a special
SWF file that works as the player and only add special code to set the decoder
stream on that SWF.