Add a simple pulseaudio playback backend.
It appears to be playing on par with the OSS backend for absinthe.swf.  Both
suffer from issues with streams not ending when they should in the file.  It
also successfully plays on my PA-eanbled Linux machine, while the ALSA
backend fails (despite other ALSA programs working).

The priority order for auto config is alsa, pulse, oss.  If common desktops
move to PA, we may want to revisit this ordering.