add swfdec_as_function_call_no_preload () to allow doing stuff before preloading
use this by swfdec_as_function_call () to create the super object. And don't
create the super object in swfdec_as_frame_preload () anymore. It seems existance
of a super object depends on how a function was called, so we need to make
creation non-automatic.

Currently, various tests will break.