initial stubs for ExternalInterface
The SwfdecPlayer class got a new sparkling "scipting" property of type
SwfdecPlayerScripting which is a GObject that is public and the user is supposed
to implement it if he wants to implement ExternalInterface.
By default this is set to NULL which means ExternalInterface is not available.

Also implement the _escapeXML, _unescapeXML and _jsQuoteString functions