get rid of the slow loader
The current loader design does not really allow writing a proxy loader the
way it was written. And since the slow loader does the wrong thing technically
(It should cnfigure the bandwidth, not the loading duration) I decided to get
rid of it.
If anyone wants to write a new one, It should probably implement the
SwfdecStreamTarget interface.