update GStreamer requirements for missing plugin stuff
Yes, it requires GStreamer 0.10.17 now (actually 0.10.15 is fine, but would
bloat configure.ac). It's a bit unfortunate, just like libsoup-2.4, but should
be solved with the next round of distros. Hardy, unstable and rawhide pretty
likely have packages for this. Hardy does, and it's easy to just grab them for
Gutsy.