add how swfdec_player_advance() works
It used to tigger all the timeouts that were necessary to advance the specified
msecs. Now it just advances to the first event. This was necessary to allow
easy abort on events like missing plugins.

Ilike that it has surprisingly little impact on the code we have already
written, because that one advances by swfdec_player_get_next_event()'s return
value, which is exactly what this change enforces. I guess the only app not
doing this will be the thumbnailer, and I haven't looked at it yet.