make StreamTargetClass.parse return a boolean
TRUE = parsed stuff, please call this function again
FALSE = done parsing

So FALSE is the same as the behavior was previously. All functions were made to
use it.
TRUE is only used by NetStream so far, which returns TRUE after parsing an
initial video/audio packet, after which it exits to allow the plugin finding
algorithms to find the right plugin. Tested to work with Youtube.