add a simple tool for changing Flash files
The goal is to have a list of possible actions that can be performed on the
file, like "add a trace to the beginning of every function" name or "remove all
traces of the character 56". This should be done without modifying other
aspects of the Flash file to preserve bugs.