fix crash when color-transforming masks
As the mask flag is used to indicate an image surface, cached masks caused
crashes, as they might not result in image surfaces.
The code was changed to just not render masks but return a black pattern
instead.

I also have a test for this. But the test doesn't fail in the testsuite, as
the testsuite only uses image surfaces. (Booo!)