Allow object class's get vfunc have NULL as value, and add ignore_pboject
When value == NULL, the steps needed to get the value should be skipped. For
example running the get function for addProperty variables.

ignore_pobject is used to tell that certain variable should be considered to be
part of the original object even if it's in the __proto__. This is the case for properties added with addProperty.