revert 15f46394467fc94a7b142700e8d65ed14eb24aca
There is multiple websites that stopped emitting onLoad events when the
MovieClip's base class (the one registered via Object.registerClass) had an
onLoad function set.

I did not revert the test, so we have a failure now.

I assume we need a better general method to detect if we should or should not
queue events. It's certainly not the if (has_clip_events) method.