make numbers garbage-collected objects
This is the first step in a reorganization of the script engine to fix
bugs that surfaced during the end of the 0.8 development cycle.
Note that most of these fixes are from going
SWFDEC_AS_VALUE_SET_NUMBER() to swfdec_as_value_set_number(), which
takes the context as an argument.