redo socket writing API
previously the SwfdecSocket was responsible for keeping a queue of data
to be written.
Now the queue is supposed to be kept inside Swfdec. This has 2
advantages:
a) The API maps nicely to async I/O, the code in external apps can be
   reduced.
b) The RTMP code can assemble data on demand and does not have to queue
   all of the data in random order inside the socket.
This will liekly only effect the written data on clogged networks.